+++ bmp_file_handler.h
@@ -99,7 +99,7 @@ typedef struct {
 } Color;
 
 // Function prototypes
-uint32_t pad_width(int32_t width, uint16_t bit_depth);
+uint32_t pad_width(int32_t width, uint8_t bit_depth);
 char *create_filename_with_suffix(char *filename, char *suffix);
 void init_bitmap(Bitmap *bmp );
 int load_bitmap(Bitmap *bmp, char *filename);
